power: move temperature calculations to api
